@@ -9,17 +9,19 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>ramzet89</groupId>
-    <artifactId>diictionary</artifactId>
+    <artifactId>dictionary</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>diictionary</name>
-    <description>Demo project for Spring Boot</description>
+    <name>dictionary</name>
+    <description>Dictionary trainer</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
+        <testcontainers.version>1.15.2</testcontainers.version>
+        <liquibase.version>4.3.1</liquibase.version>
     </properties>
 
     <dependencies>
-<!--        spring-->
+        <!--        spring-->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
@@ -35,13 +37,13 @@
             <artifactId>spring-boot-starter-security</artifactId>
         </dependency>
 
-<!--        security-->
+        <!--        security-->
         <dependency>
             <groupId>io.jsonwebtoken</groupId>
             <artifactId>jjwt</artifactId>
             <version>0.9.1</version>
         </dependency>
-<!--        lombok-->
+        <!--        lombok-->
 
         <dependency>
             <groupId>org.projectlombok</groupId>
@@ -50,7 +52,7 @@
             <scope>provided</scope>
         </dependency>
 
-<!--        additional-->
+        <!--        additional-->
 
         <dependency>
             <groupId>org.jetbrains</groupId>
@@ -64,8 +66,6 @@
             <version>2.0.1.Final</version>
         </dependency>
 
-
-
         <!--        Test-->
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -77,6 +77,20 @@
                     <artifactId>junit-vintage-engine</artifactId>
                 </exclusion>
             </exclusions>
+        </dependency>
+
+        <dependency>
+            <groupId>org.testcontainers</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>${testcontainers.version}</version>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.testcontainers</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <version>${testcontainers.version}</version>
+            <scope>test</scope>
         </dependency>
 
         <!--        serialize-->
@@ -95,12 +109,12 @@
         <dependency>
             <groupId>org.liquibase</groupId>
             <artifactId>liquibase-core</artifactId>
-            <version>4.1.1</version>
+            <version>${liquibase.version}</version>
         </dependency>
         <dependency>
             <groupId>org.liquibase</groupId>
             <artifactId>liquibase-maven-plugin</artifactId>
-            <version>4.1.1</version>
+            <version>${liquibase.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -122,7 +136,7 @@
                     <execution>
                         <phase>process-resources</phase>
                         <goals>
-                            <goal>update</goal>
+                            <!--                            <goal>update</goal>-->
                         </goals>
                     </execution>
                 </executions>
@@ -132,8 +146,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.5.1</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>${java.version}</source>
+                    <target>${java.version}</target>
                     <annotationProcessorPaths>
                         <path>
                             <groupId>org.mapstruct</groupId>
