@@ -5,8 +5,6 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
-import java.util.Arrays;
-import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -15,8 +13,8 @@ import static ramzet89.dictionary.enums.security.Permission.*;
 @Getter
 @RequiredArgsConstructor
 public enum UserRole {
-    ADMIN(new HashSet<>(Arrays.asList(ADMIN_USERS, USER_LEARN, USER_ADD))),
-    USER(new HashSet<>(Arrays.asList(USER_ADD, USER_LEARN)));
+    ADMIN(Set.of(ADMIN_USERS, USER_LEARN, USER_ADD)),
+    USER(Set.of(USER_ADD, USER_LEARN));
 
     private final Set<Permission> permissions;
 
